refactor: alterado nome da pasta "controller" para "menu"
