+++ agenda/src/menu/Menu.java
@@ -1,4 +1,4 @@
-package controller;
+package menu;
 
 import java.util.List;
 import java.util.Scanner;
