+++ agenda/src/menu/MenuEventos.java
@@ -1,4 +1,4 @@
-package controller;
+package menu;
 import model.Evento;
 import service.agenda_contatos.AgendaEventosService;
 import java.time.LocalDate;
