+++ agenda/src/menu/MenuWheater.java
@@ -1,4 +1,4 @@
-package controller;
+package menu;
 
 import view.WeatherForecast;
 import util.Util;
