+++ agenda/src/service/delivery/MenuDeliveryService.java
@@ -3,7 +3,7 @@
 
 package service.delivery;
 
-import controller.Menu;
+import menu.Menu;
 import model.ItemService;
 import util.Util;
