+++ agenda/src/service/delivery/SubmenuService.java
@@ -2,7 +2,7 @@
 
 package service.delivery;
 
-import controller.Menu;
+import menu.Menu;
 import model.ItemService;
 import util.Util;
